[BEAM-7862] Moving FakeBigQueryServices to published artifacts (#9206)
* Moving FakeBigQueryServices to published artifacts

* Fixing ApiSurface tests
